Move landmarks into a submodule
