+++ MANIFEST.in
@@ -1,3 +1,3 @@
 include LICENSE
 include requirements.txt
-exclude */test_*.py
+exclude **/test_*.py
