+++ dev.py
@@ -21,6 +21,11 @@ def init():
     execute("pip install --upgrade -r requirements_dev.txt")
 
 
+@cli.command()
+def clean():
+    execute("find . -name '*.pyc' -delete")
+
+
 def docker_repo(python_version, tag):
     return "laceproject/entente-ci-{}:{}".format(python_version, tag)
 
@@ -56,12 +61,12 @@ def docker_push(tag):
 
 @cli.command()
 def test():
-    execute("python -m pytest")
+    execute("python3 -m pytest")
 
 
 @cli.command()
 def coverage():
-    execute("python -m pytest --cov=entente")
+    execute("python3 -m pytest --cov=entente")
 
 
 @cli.command()
