+++ entente/landmarks/landmarker.py
@@ -91,7 +91,7 @@ class Landmarker(object):
         Returns:
             dict: A mapping of landmark names to a np.ndarray with shape `3x1`.
         """
-        from .equality import have_same_topology
+        from ..equality import have_same_topology
 
         if not have_same_topology(self.source_mesh, target):
             raise ValueError("Target mesh must have the same topology")
